feat: improve TypeScript support (#18)
* feat: improve TypeScript support

* fix: use tsx instead of ts-node to run code

* fix: update pnpm-lock.yaml

* fix: remove unused code

* fix: handle incomplete SSE  stream as error

---------
