+++ examples/coze-js-node/src/auth/auth-oauth-jwt-channel.ts
@@ -1,7 +1,6 @@
-/* eslint-disable no-unused-vars */
-
+/* eslint-disable @typescript-eslint/naming-convention */
 /*
- * How to effectuate OpenAPI authorization through the OAuth JWT (JSON Web Token) method.
+ * How to effectuate OpenAPI authorization through the OAuth JWT (JSON Web Token) method for channel access.
  *
  * Firstly, users need to access https://www.coze.com/open/oauth/apps. For the cn environment,
  * users need to access https://www.coze.cn/open/oauth/apps to create an OAuth App of the type
@@ -20,24 +19,25 @@ import { dirname, join } from 'node:path';
 import fs from 'fs';
 
 import jwt from 'jsonwebtoken';
-import { CozeAPI, getJWTToken } from '@coze/api';
+import { CozeAPI, getJWTToken, RoleType } from '@coze/api';
 
-import config from '../../config.js';
+import config from '../config/config';
 
 // 'en' for https://api.coze.com, 'cn' for https://api.coze.cn
 const key = process.env.COZE_ENV || 'en';
 
 // Retrieve configuration values from the config file
 const baseURL = config[key].COZE_BASE_URL;
-const appId = config[key].auth.oauth_jwt.COZE_APP_ID;
-const keyid = config[key].auth.oauth_jwt.COZE_KEY_ID;
-const aud = config[key].auth.oauth_jwt.COZE_AUD;
+const appId = config[key].auth.auth_jwt_channel.COZE_APP_ID;
+const keyid = config[key].auth.auth_jwt_channel.COZE_KEY_ID;
+const aud = config[key].auth.auth_jwt_channel.COZE_AUD;
+const botIdList = [config[key].COZE_BOT_ID];
 
 // Read the private key from a file
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 const privateKey = fs.readFileSync(
-  join(__dirname, '../../tmp/private_key.pem'),
+  join(__dirname, '../../tmp/private_key2.pem'),
 );
 
 // Prepare the payload for the JWT
@@ -56,20 +56,44 @@ jwt.sign(
   privateKey,
   { algorithm: 'RS256', keyid },
   async (err, token) => {
-    console.log('token', token);
+    console.log('jwtToken', token);
 
-    // Exchange the JWT for an OAuth token
+    // Define the scope for channel access
+    const scope = {
+      account_permission: {
+        permission_list: ['Connector.botChat'],
+      },
+      attribute_constraint: {
+        connector_bot_chat_attribute: {
+          bot_id_list: botIdList,
+        },
+      },
+    };
+
+    // Exchange the JWT for an OAuth token with specified scope
     const result = await getJWTToken({
       baseURL,
       token,
       duration_seconds: 86399,
+      scope,
     });
     console.log('getJWTToken', result);
 
     // Initialize a new Coze API client using the obtained access token
     const client = new CozeAPI({ baseURL, token: result.access_token });
 
-    // Example of how to use the client (commented out)
-    // e.g. client.chat.stream(...);
+    // Example of how to use the client for chat creation
+    const data = await client.chat.create({
+      bot_id: botIdList[0],
+      user_id: '1234567890',
+      additional_messages: [
+        {
+          role: RoleType.User,
+          content: 'Hello',
+          content_type: 'text',
+        },
+      ],
+    });
+    console.log('client.chat.create', data);
   },
 );
