+++ examples/coze-js-node/src/chat.ts
@@ -1,28 +1,30 @@
-/* eslint-disable no-unused-vars */
-import { clearLine, cursorTo } from 'node:readline';
 import assert from 'assert';
 
-import { ChatEventType, ChatStatus } from '@coze/api';
+import {
+  ChatEventType,
+  ChatStatus,
+  type CreateChatData,
+  RoleType,
+} from '@coze/api';
 
-import { client, botId, sleep } from './client.mjs';
+import { client, botId, sleep } from './client';
 
-const query = '来一段有趣的代码';
+const query = 'give me a joke';
 
-async function streamingChat(callback) {
+async function streamingChat(callback?: (v: CreateChatData) => void) {
   assert(botId, 'botId is required');
   const v = await client.chat.stream({
     bot_id: botId,
-    user_id: '123',
     auto_save_history: false,
     additional_messages: [
       {
-        role: 'user',
+        role: RoleType.User,
         content: query,
         content_type: 'text',
       },
     ],
   });
-  console.log('xx', v);
+
   for await (const part of v) {
     if (part.event === ChatEventType.CONVERSATION_CHAT_CREATED) {
       console.log('[START]');
@@ -40,6 +42,8 @@ async function streamingChat(callback) {
       console.log(part.data.usage);
     } else if (part.event === ChatEventType.DONE) {
       console.log(part.data);
+    } else if (part.event === ChatEventType.ERROR) {
+      console.error(part.data);
     }
   }
 
@@ -51,17 +55,16 @@ async function nonStreamingChat() {
 
   const v = await client.chat.createAndPoll({
     bot_id: botId,
-    user_id: '123',
     additional_messages: [
       {
-        role: 'user',
+        role: RoleType.User,
         content: query,
         content_type: 'text',
       },
     ],
   });
   if (v.chat.status === ChatStatus.COMPLETED) {
-    for (const item of v.messages) {
+    for (const item of v.messages || []) {
       console.log('[%s]:[%s]:%s', item.role, item.type, item.content);
     }
     console.log('usage', v.chat.usage);
@@ -69,17 +72,18 @@ async function nonStreamingChat() {
 }
 
 // TODO  test
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
 async function submitToolOutputs() {
   const v = await client.chat.submitToolOutputs({
     conversation_id: '123',
     chat_id: '123',
-    tool_outputs: [{ id: '123', output: '123' }],
+    tool_outputs: [{ tool_call_id: '123', output: '123' }],
     stream: false,
   });
   console.log('client.chat.submitToolOutputs', v);
 }
 
-async function streamingCancel() {
+function streamingCancel() {
   streamingChat(async v => {
     sleep(2000);
     const result = await client.chat.cancel(v.conversation_id, v.id);
