+++ examples/coze-js-node/src/voice.ts
@@ -1,8 +1,12 @@
+/* eslint-disable @typescript-eslint/naming-convention */
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import { fileURLToPath } from 'node:url';
 import { dirname, join } from 'node:path';
 import fs from 'fs';
 
-import { client } from './client.mjs';
+import { type Voice } from '@coze/api';
+
+import { client } from './client';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
@@ -10,21 +14,19 @@ const __dirname = dirname(__filename);
 const filePath = join(__dirname, '../tmp/voice.mp3');
 const fileBuffer = await fs.createReadStream(filePath);
 
-try {
+async function voiceClone() {
   const voiceObj = await client.audio.voices.clone({
     audio_format: 'mp3',
-    file: fileBuffer,
+    file: fileBuffer as any,
     voice_name: '湾湾小何2',
     preview_text:
       '今天天气真是太好了，阳光灿烂，心情超级棒，但是朋友最近的感情问题也让我心痛不已，好像世界末日一样，真的好为他难过。',
     voice_id: '742894224871836***',
   });
   console.log('client.audio.voices.clone', voiceObj);
-} catch (error) {
-  console.error('Failed to clone voice:', error.message);
 }
 
-try {
+async function createSpeech() {
   const speechBuffer = await client.audio.speech.create({
     input:
       '今天天气真是太好了，阳光灿烂，心情超级棒，但是朋友最近的感情问题也让我心痛不已，好像世界末日一样，真的好为他难过。',
@@ -37,35 +39,31 @@ try {
     console.warn(`Warning: Overwriting existing file at ${audioPath}`);
   }
 
-  await fs.writeFileSync(audioPath, speechBuffer, 'binary');
+  await fs.writeFileSync(audioPath, speechBuffer as any, 'binary');
   console.log('Speech saved successfully:', audioPath);
-} catch (error) {
-  console.error('Failed to generate or save speech:', error.message);
 }
 
 async function listAllVoices() {
-  try {
-    const PAGE_SIZE = 20;
-    let page = 1;
-    let allVoices = [];
+  const PAGE_SIZE = 20;
+  let page = 1;
+  let allVoices: Voice[] = [];
 
-    while (true) {
-      const response = await client.audio.voices.list({
-        page_size: PAGE_SIZE,
-        page_num: page,
-      });
-      allVoices = allVoices.concat(response.voice_list);
-      if (response.has_more === false) {
-        break;
-      }
-      page++;
+  while (true) {
+    const response = await client.audio.voices.list({
+      page_size: PAGE_SIZE,
+      page_num: page,
+    });
+    allVoices = allVoices.concat(response.voice_list);
+    if (response.has_more === false) {
+      break;
     }
-
-    console.log(`Total voices found: ${allVoices.length}`);
-    return allVoices;
-  } catch (error) {
-    console.error('Failed to list voices:', error.message);
+    page++;
   }
+
+  console.log(`Total voices found: ${allVoices.length}`);
+  return allVoices;
 }
 
-await listAllVoices();
+await voiceClone().catch(console.error);
+await createSpeech().catch(console.error);
+await listAllVoices().catch(console.error);
