+++ examples/coze-js-node/src/workspace.ts
@@ -1,8 +1,8 @@
-import { client } from './client.mjs';
+import { client } from './client';
 
 async function main() {
   const list = await client.workspaces.list({ page_num: 1, page_size: 10 });
   console.log('client.workspaces.list', list);
 }
 
-main();
+main().catch(console.error);
