+++ packages/coze-js/src/resources/chat/chat.ts
@@ -463,7 +463,7 @@ export interface StreamChatReq {
   /**
    * 标识当前与 Bot 交互的用户，由使用方在业务系统中自行定义、生成与维护。
    */
-  user_id: string;
+  user_id?: string;
 
   /**
    * 对话的附加信息。你可以通过此字段传入本次对话中用户的问题。数组长度限制为 100，即最多传入 100 条消息。
